Change ReindexJobCompleted from mediatr message to ReindexUtilitites
