--- src/Microsoft.Health.Fhir.Core/Messages/Reindex/ReindexJobCompletedResponse.cs
@@ -1,20 +1,0 @@
-﻿// -------------------------------------------------------------------------------------------------
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License (MIT). See LICENSE in the repo root for license information.
-// -------------------------------------------------------------------------------------------------
-
-namespace Microsoft.Health.Fhir.Core.Messages.Reindex
-{
-    public class ReindexJobCompletedResponse
-    {
-        public ReindexJobCompletedResponse(bool success, string errorMessage)
-        {
-            Success = success;
-            ErrorMessage = errorMessage;
-        }
-
-        public bool Success { get; }
-
-        public string ErrorMessage { get; }
-    }
-}
